Convert checkRaces to ts (#464)
* Convert checkRaces to TS

* Refactoring
